requirements.txt
gdown==5.2.0
numpy==2.2.4
Pillow==11.1.0
streamlit==1.43.2
tensorflow==2.19.0
